netlist: add global added cell cache.
This replaces the cell cache previously used in the decision pass.

Maintaining a hashmap of all added cells is very cheap and it can save
a significant amount of additional work (before the cells are removed
by the merge pass, simplify will operate on all of them). Also, too
many identical cells make IR output undebuggable without processing,
which changes cell indices.

The cell cache is cleared after every `apply()` call. Since the keys in
the cache are be mutated when nets are replaced, it would have to be
cleared before rebuilding if it was persisted across `apply()` calls.