netlist: use [] instead of {} for concatenations.
We already use {} for denoting `match` and `target` cell contents,
which is syntactically and semantically unrelated, and will create more
issues if we try to make `target` cell printing more concise.

Also, use () instead of [] for alternate groups in `match` cells as
a result, since those would otherwise be confusable with concatenations.