yosys_json: do not automatically export target cells.
Exporting target cells disturbs cell numbering, which is the primary
way of cross-referencing uir and json files and timing reports.

To allow such cross-referencing, exporting is now done solely in
the CLI and a new `--export`/`-e` option enables it for `.uir` files.
For `.json` files this is done unconditionally.